Relocate ProcessIO to ContainerClient.
- Makes ProcessIO more generally available.
- `Application.handleProcess(processIO:process:)` becomes
  `processIO.handleProcess(process:log:)`.
