+++ Sources/ContainerCommands/Container/ContainerRun.swift
@@ -154,7 +154,7 @@ extension Application {
                     }
                 }
 
-                exitCode = try await Application.handleProcess(io: io, process: process)
+                exitCode = try await io.handleProcess(process: process, log: log)
             } catch {
                 if error is ContainerizationError {
                     throw error
@@ -162,288 +162,6 @@ extension Application {
                 throw ContainerizationError(.internalError, message: "failed to run container: \(error)")
             }
             throw ArgumentParser.ExitCode(exitCode)
-        }
-    }
-}
-
-public struct ProcessIO: Sendable {
-    let stdin: Pipe?
-    let stdout: Pipe?
-    let stderr: Pipe?
-    var ioTracker: IoTracker?
-
-    public struct IoTracker: Sendable {
-        let stream: AsyncStream<Void>
-        let cont: AsyncStream<Void>.Continuation
-        let configuredStreams: Int
-    }
-
-    let stdio: [FileHandle?]
-
-    let console: Terminal?
-
-    func closeAfterStart() throws {
-        try stdin?.fileHandleForReading.close()
-        try stdout?.fileHandleForWriting.close()
-        try stderr?.fileHandleForWriting.close()
-    }
-
-    func close() throws {
-        try console?.reset()
-    }
-
-    static func create(tty: Bool, interactive: Bool, detach: Bool) throws -> ProcessIO {
-        let current: Terminal? = try {
-            if !tty || !interactive {
-                return nil
-            }
-            let current = try Terminal.current
-            try current.setraw()
-            return current
-        }()
-
-        var stdio = [FileHandle?](repeating: nil, count: 3)
-
-        let stdin: Pipe? = {
-            if !interactive {
-                return nil
-            }
-            return Pipe()
-        }()
-
-        if let stdin {
-            let pin = FileHandle.standardInput
-            let stdinOSFile = OSFile(fd: pin.fileDescriptor)
-            let pipeOSFile = OSFile(fd: stdin.fileHandleForWriting.fileDescriptor)
-            try stdinOSFile.makeNonBlocking()
-            nonisolated(unsafe) let buf = UnsafeMutableBufferPointer<UInt8>.allocate(capacity: Int(getpagesize()))
-
-            pin.readabilityHandler = { _ in
-                Self.streamStdin(
-                    from: stdinOSFile,
-                    to: pipeOSFile,
-                    buffer: buf,
-                ) {
-                    pin.readabilityHandler = nil
-                    buf.deallocate()
-                    try? stdin.fileHandleForWriting.close()
-                }
-            }
-            stdio[0] = stdin.fileHandleForReading
-        }
-
-        let stdout: Pipe? = {
-            if detach {
-                return nil
-            }
-            return Pipe()
-        }()
-
-        var configuredStreams = 0
-        let (stream, cc) = AsyncStream<Void>.makeStream()
-        if let stdout {
-            configuredStreams += 1
-            let pout: FileHandle = {
-                if let current {
-                    return current.handle
-                }
-                return .standardOutput
-            }()
-
-            let rout = stdout.fileHandleForReading
-            rout.readabilityHandler = { handle in
-                let data = handle.availableData
-                if data.isEmpty {
-                    rout.readabilityHandler = nil
-                    cc.yield()
-                    return
-                }
-                try! pout.write(contentsOf: data)
-            }
-            stdio[1] = stdout.fileHandleForWriting
-        }
-
-        let stderr: Pipe? = {
-            if detach || tty {
-                return nil
-            }
-            return Pipe()
-        }()
-        if let stderr {
-            configuredStreams += 1
-            let perr: FileHandle = .standardError
-            let rerr = stderr.fileHandleForReading
-            rerr.readabilityHandler = { handle in
-                let data = handle.availableData
-                if data.isEmpty {
-                    rerr.readabilityHandler = nil
-                    cc.yield()
-                    return
-                }
-                try! perr.write(contentsOf: data)
-            }
-            stdio[2] = stderr.fileHandleForWriting
-        }
-
-        var ioTracker: IoTracker? = nil
-        if configuredStreams > 0 {
-            ioTracker = .init(stream: stream, cont: cc, configuredStreams: configuredStreams)
-        }
-
-        return .init(
-            stdin: stdin,
-            stdout: stdout,
-            stderr: stderr,
-            ioTracker: ioTracker,
-            stdio: stdio,
-            console: current
-        )
-    }
-
-    static func streamStdin(
-        from: OSFile,
-        to: OSFile,
-        buffer: UnsafeMutableBufferPointer<UInt8>,
-        onErrorOrEOF: () -> Void,
-    ) {
-        while true {
-            let (bytesRead, action) = from.read(buffer)
-            if bytesRead > 0 {
-                let view = UnsafeMutableBufferPointer(
-                    start: buffer.baseAddress,
-                    count: bytesRead
-                )
-
-                let (bytesWritten, _) = to.write(view)
-                if bytesWritten != bytesRead {
-                    onErrorOrEOF()
-                    return
-                }
-            }
-
-            switch action {
-            case .error(_), .eof, .brokenPipe:
-                onErrorOrEOF()
-                return
-            case .again:
-                return
-            case .success:
-                break
-            }
-        }
-    }
-
-    func wait() async throws {
-        guard let ioTracker = self.ioTracker else {
-            return
-        }
-        do {
-            try await Timeout.run(seconds: 3) {
-                var counter = ioTracker.configuredStreams
-                for await _ in ioTracker.stream {
-                    counter -= 1
-                    if counter == 0 {
-                        ioTracker.cont.finish()
-                        break
-                    }
-                }
-            }
-        } catch {
-            log.error("Timeout waiting for IO to complete : \(error)")
-            throw error
-        }
-    }
-}
-
-public struct OSFile: Sendable {
-    private let fd: Int32
-
-    public enum IOAction: Equatable {
-        case eof
-        case again
-        case success
-        case brokenPipe
-        case error(_ errno: Int32)
-    }
-
-    public init(fd: Int32) {
-        self.fd = fd
-    }
-
-    public init(handle: FileHandle) {
-        self.fd = handle.fileDescriptor
-    }
-
-    func makeNonBlocking() throws {
-        let flags = fcntl(fd, F_GETFL)
-        guard flags != -1 else {
-            throw POSIXError.fromErrno()
-        }
-
-        if fcntl(fd, F_SETFL, flags | O_NONBLOCK) == -1 {
-            throw POSIXError.fromErrno()
-        }
-    }
-
-    func write(_ buffer: UnsafeMutableBufferPointer<UInt8>) -> (wrote: Int, action: IOAction) {
-        if buffer.count == 0 {
-            return (0, .success)
-        }
-
-        var bytesWrote: Int = 0
-        while true {
-            let n = Darwin.write(
-                self.fd,
-                buffer.baseAddress!.advanced(by: bytesWrote),
-                buffer.count - bytesWrote
-            )
-            if n == -1 {
-                if errno == EAGAIN || errno == EIO {
-                    return (bytesWrote, .again)
-                }
-                return (bytesWrote, .error(errno))
-            }
-
-            if n == 0 {
-                return (bytesWrote, .brokenPipe)
-            }
-
-            bytesWrote += n
-            if bytesWrote < buffer.count {
-                continue
-            }
-            return (bytesWrote, .success)
-        }
-    }
-
-    func read(_ buffer: UnsafeMutableBufferPointer<UInt8>) -> (read: Int, action: IOAction) {
-        if buffer.count == 0 {
-            return (0, .success)
-        }
-
-        var bytesRead: Int = 0
-        while true {
-            let n = Darwin.read(
-                self.fd,
-                buffer.baseAddress!.advanced(by: bytesRead),
-                buffer.count - bytesRead
-            )
-            if n == -1 {
-                if errno == EAGAIN || errno == EIO {
-                    return (bytesRead, .again)
-                }
-                return (bytesRead, .error(errno))
-            }
-
-            if n == 0 {
-                return (bytesRead, .eof)
-            }
-
-            bytesRead += n
-            if bytesRead < buffer.count {
-                continue
-            }
-            return (bytesRead, .success)
         }
     }
 }
